Remove process spawn, old TypeScript versions
