+++ .travis.yml
@@ -15,12 +15,8 @@ script:
   - npm run test-cov
 
 env:
-  - NODE=8 TYPESCRIPT=typescript@1.6
-  - NODE=8 TYPESCRIPT=typescript@1.7
-  - NODE=8 TYPESCRIPT=typescript@1.8
   - NODE=8 TYPESCRIPT=typescript@2.0
-  - NODE=8 TYPESCRIPT=typescript@2.1
-  - NODE=4 TYPESCRIPT=typescript@latest
+  - NODE=6 TYPESCRIPT=typescript@latest
   - NODE=8 TYPESCRIPT=typescript@latest
   - NODE=8 TYPESCRIPT=typescript@next
