+++ package.json
@@ -5,8 +5,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "bin": {
-    "ts-node": "dist/bin.js",
-    "_ts-node": "dist/_bin.js"
+    "ts-node": "dist/bin.js"
   },
   "files": [
     "dist/",
@@ -14,7 +13,7 @@
     "LICENSE"
   ],
   "scripts": {
-    "lint": "tslint \"src/**/*.ts\" --project tsconfig.json --type-check",
+    "lint": "tslint \"src/**/*.ts\" --project tsconfig.json",
     "clean": "rimraf dist",
     "tsc": "tsc",
     "build": "npm run clean && npm run tsc",
@@ -50,21 +49,20 @@
   "homepage": "https://github.com/TypeStrong/ts-node",
   "devDependencies": {
     "@types/arrify": "^1.0.1",
+    "@types/chai": "^4.0.4",
     "@types/diff": "^3.2.1",
     "@types/minimist": "^1.2.0",
     "@types/mkdirp": "^0.5.0",
-    "@types/node": "^8.0.27",
-    "@types/source-map-support": "^0.4.0",
-    "@types/v8flags": "types/npm-v8flags#de224ae1cd5fd7dbb4e7158a6cc7a29e5315930d",
-    "@types/yn": "types/npm-yn#ca75f6c82940fae6a06fb41d2d37a6aa9b4ea8e9",
-    "@types/chai": "^4.0.4",
     "@types/mocha": "^2.2.42",
+    "@types/node": "^9.4.6",
     "@types/proxyquire": "^1.3.28",
     "@types/react": "^16.0.2",
     "@types/semver": "^5.3.34",
+    "@types/source-map-support": "^0.4.0",
+    "@types/yn": "types/npm-yn#ca75f6c82940fae6a06fb41d2d37a6aa9b4ea8e9",
     "chai": "^4.0.1",
     "istanbul": "^0.4.0",
-    "mocha": "^4.0.0",
+    "mocha": "^5.0.1",
     "ntypescript": "^1.201507091536.1",
     "proxyquire": "^1.7.2",
     "react": "^16.0.0",
@@ -72,7 +70,7 @@
     "semver": "^5.1.0",
     "tslint": "^5.0.0",
     "tslint-config-standard": "^7.0.0",
-    "typescript": "^2.4.1"
+    "typescript": "^2.7.2"
   },
   "dependencies": {
     "arrify": "^1.0.0",
@@ -81,9 +79,7 @@
     "make-error": "^1.1.1",
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
-    "source-map-support": "^0.5.0",
-    "tsconfig": "^7.0.0",
-    "v8flags": "^3.0.0",
+    "source-map-support": "^0.5.3",
     "yn": "^2.0.0"
   }
 }
