+++ src/index.ts
@@ -1,5 +1,5 @@
 import { relative, basename, extname, resolve, dirname, join } from 'path'
-import { writeFileSync, readFileSync, statSync } from 'fs'
+import { readFileSync, writeFileSync } from 'fs'
 import { EOL, tmpdir, homedir } from 'os'
 import sourceMapSupport = require('source-map-support')
 import chalk from 'chalk'
@@ -9,7 +9,6 @@ import yn = require('yn')
 import arrify = require('arrify')
 import { BaseError } from 'make-error'
 import * as TS from 'typescript'
-import { loadSync } from 'tsconfig'
 
 const pkg = require('../package.json')
 const shouldDebug = yn(process.env.TS_NODE_DEBUG)
@@ -41,9 +40,6 @@ export interface TSCommon {
   findConfigFile: typeof TS.findConfigFile
   readConfigFile: typeof TS.readConfigFile
   parseJsonConfigFileContent: typeof TS.parseJsonConfigFileContent
-
-  // TypeScript 1.5 and 1.6.
-  parseConfigFile? (json: any, host: any, basePath: string): any
 }
 
 /**
@@ -55,23 +51,25 @@ export const VERSION = pkg.version
  * Registration options.
  */
 export interface Options {
-  typeCheck?: boolean
-  cache?: boolean
+  typeCheck?: boolean | null
+  cache?: boolean | null
   cacheDirectory?: string
   compiler?: string
-  project?: boolean | string
-  ignore?: boolean | string | string[]
-  ignoreWarnings?: number | string | Array<number | string>
-  getFile?: (path: string) => string
+  ignore?: string | string[]
+  project?: string
+  skipIgnore?: boolean | null
+  skipProject?: boolean | null
+  compilerOptions?: object
+  ignoreDiagnostics?: number | string | Array<number | string>
+  readFile?: (path: string) => string | undefined
   fileExists?: (path: string) => boolean
-  compilerOptions?: any
   transformers?: TS.CustomTransformers
 }
 
 /**
  * Track the project information.
  */
-interface Cache {
+interface MemoryCache {
   contents: { [path: string]: string }
   versions: { [path: string]: number | undefined }
   outputs: { [path: string]: string }
@@ -88,31 +86,43 @@ export interface TypeInfo {
 /**
  * Default register options.
  */
-const DEFAULTS = {
-  getFile,
-  fileExists,
+export const DEFAULTS: Options = {
   cache: yn(process.env['TS_NODE_CACHE'], { default: true }),
   cacheDirectory: process.env['TS_NODE_CACHE_DIRECTORY'],
   compiler: process.env['TS_NODE_COMPILER'],
   compilerOptions: parse(process.env['TS_NODE_COMPILER_OPTIONS']),
-  project: process.env['TS_NODE_PROJECT'],
   ignore: split(process.env['TS_NODE_IGNORE']),
-  ignoreWarnings: split(process.env['TS_NODE_IGNORE_WARNINGS']),
+  project: process.env['TS_NODE_PROJECT'],
+  skipIgnore: yn(process.env['TS_NODE_SKIP_IGNORE']),
+  skipProject: yn(process.env['TS_NODE_SKIP_PROJECT']),
+  ignoreDiagnostics: split(process.env['TS_NODE_IGNORE_DIAGNOSTICS']),
   typeCheck: yn(process.env['TS_NODE_TYPE_CHECK'])
+}
+
+/**
+ * Default TypeScript compiler options required by `ts-node`.
+ */
+const DEFAULT_COMPILER_OPTIONS = {
+  sourceMap: true,
+  inlineSourceMap: false,
+  inlineSources: true,
+  declaration: false,
+  noEmit: false,
+  outDir: '$$ts-node$$'
 }
 
 /**
  * Split a string array of values.
  */
 export function split (value: string | undefined) {
-  return value ? value.split(/ *, */g) : undefined
+  return typeof value === 'string' ? value.split(/ *, */g) : undefined
 }
 
 /**
  * Parse a string as JSON.
  */
-export function parse (value: string | undefined) {
-  return value ? JSON.parse(value) : undefined
+export function parse (value: string | undefined): object | undefined {
+  return typeof value === 'string' ? JSON.parse(value) : undefined
 }
 
 /**
@@ -126,7 +136,6 @@ export function normalizeSlashes (value: string): string {
  * TypeScript diagnostics error.
  */
 export class TSError extends BaseError {
-
   name = 'TSError'
 
   constructor (public diagnostics: TSDiagnostic[]) {
@@ -134,7 +143,6 @@ export class TSError extends BaseError {
       `⨯ Unable to compile TypeScript\n${diagnostics.map(x => x.message).join('\n')}`
     )
   }
-
 }
 
 /**
@@ -161,54 +169,46 @@ function getTmpDir (): string {
 /**
  * Register TypeScript compiler.
  */
-export function register (options: Options = {}): Register {
-  const compiler = options.compiler || 'typescript'
+export function register (opts: Options = {}): Register {
   const emptyFileListWarnings = [18002, 18003]
-  const ignoreWarnings = arrify(
-    options.ignoreWarnings || DEFAULTS.ignoreWarnings || []
-  ).concat(emptyFileListWarnings).map(Number)
-  const getFile = options.getFile || DEFAULTS.getFile
-  const fileExists = options.fileExists || DEFAULTS.fileExists
-  const shouldCache = !!(options.cache === undefined ? DEFAULTS.cache : options.cache)
-  const typeCheck = !!(options.typeCheck === undefined ? DEFAULTS.typeCheck : options.typeCheck)
-  const project = options.project === undefined ? DEFAULTS.project : options.project
-  const cacheDirectory = options.cacheDirectory || DEFAULTS.cacheDirectory || getTmpDir()
-  const compilerOptions = Object.assign({}, DEFAULTS.compilerOptions, options.compilerOptions)
+  const options = Object.assign({}, DEFAULTS, opts)
+  const cacheDirectory = options.cacheDirectory || getTmpDir()
+  const ignoreDiagnostics = arrify(options.ignoreDiagnostics).concat(emptyFileListWarnings).map(Number)
   const originalJsHandler = require.extensions['.js']
 
-  const cache: Cache = {
+  const memoryCache: MemoryCache = {
     contents: Object.create(null),
     versions: Object.create(null),
     outputs: Object.create(null)
   }
 
-  const ignore = arrify(
-    (
-      typeof options.ignore === 'boolean' ?
-        (options.ignore === false ? [] : undefined) :
-        (options.ignore || DEFAULTS.ignore)
-    ) ||
-    ['/node_modules/']
+  const ignore = options.skipIgnore ? [] : arrify(
+    options.ignore || '/node_modules/'
   ).map(str => new RegExp(str))
 
-  // Install source map support and read from cache.
+  // Install source map support and read from memory cache.
   sourceMapSupport.install({
     environment: 'node',
     retrieveFile (path: string) {
-      return cache.outputs[path]
+      return memoryCache.outputs[path]
     }
   })
 
   // Require the TypeScript compiler and configuration.
   const cwd = process.cwd()
+  const compiler = options.compiler || 'typescript'
+  const typeCheck = options.typeCheck || false
   const ts: typeof TS = require(compiler)
-  const config = readConfig(compilerOptions, project, cwd, ts)
-  const configDiagnostics = filterDiagnostics(config.errors, ignoreWarnings)
+  const transformers = options.transformers || undefined
+  const readFile = options.readFile || ts.sys.readFile
+  const fileExists = options.fileExists || ts.sys.fileExists
+  const config = readConfig(cwd, ts, options.compilerOptions, fileExists, readFile, options.project, options.skipProject)
+  const configDiagnostics = filterDiagnostics(config.errors, ignoreDiagnostics)
   const extensions = ['.ts', '.tsx']
 
   const cachedir = join(
     resolve(cwd, cacheDirectory),
-    getCompilerDigest({ version: ts.version, typeCheck, ignoreWarnings, config, compiler })
+    getCompilerDigest({ version: ts.version, typeCheck, ignoreDiagnostics, config, compiler })
   )
 
   // Render the configuration errors and exit the script.
@@ -224,21 +224,15 @@ export function register (options: Options = {}): Register {
 
   // Add all files into the file hash.
   for (const fileName of config.fileNames) {
-    cache.versions[fileName] = 1
+    memoryCache.versions[fileName] = 1
   }
 
   /**
    * Get the extension for a transpiled file.
    */
-  function getExtension (fileName: string) {
-    const ext = extname(fileName)
-
-    if (config.options.jsx === ts.JsxEmit.Preserve && (ext === '.tsx' || ext === '.jsx')) {
-      return '.jsx'
-    }
-
-    return '.js'
-  }
+  const getExtension = config.options.jsx === ts.JsxEmit.Preserve ?
+    ((path: string) => /\.[tj]sx$/.test(path) ? '.jsx' : '.js') :
+    ((_: string) => '.js')
 
   /**
    * Create the basic required function using transpile mode.
@@ -246,13 +240,13 @@ export function register (options: Options = {}): Register {
   let getOutput = function (code: string, fileName: string, lineOffset = 0): SourceOutput {
     const result = ts.transpileModule(code, {
       fileName,
+      transformers,
       compilerOptions: config.options,
-      reportDiagnostics: true,
-      transformers: options.transformers
+      reportDiagnostics: true
     })
 
     const diagnosticList = result.diagnostics ?
-      filterDiagnostics(result.diagnostics, ignoreWarnings) :
+      filterDiagnostics(result.diagnostics, ignoreDiagnostics) :
       []
 
     if (diagnosticList.length) {
@@ -262,15 +256,6 @@ export function register (options: Options = {}): Register {
     return [result.outputText, result.sourceMapText as string]
   }
 
-  let compile = readThrough(
-    cachedir,
-    shouldCache,
-    getFile,
-    cache,
-    getOutput,
-    getExtension
-  )
-
   let getTypeInfo = function (_code: string, _fileName: string, _position: number): TypeInfo {
     throw new TypeError(`Type information is unavailable without "--type-check"`)
   }
@@ -278,18 +263,18 @@ export function register (options: Options = {}): Register {
   // Use full language services when the fast option is disabled.
   if (typeCheck) {
     // Set the file contents into cache.
-    const setCache = function (code: string, fileName: string) {
-      if (cache.contents[fileName] !== code) {
-        cache.contents[fileName] = code
-        cache.versions[fileName] = (cache.versions[fileName] || 0) + 1
+    const updateMemoryCache = function (code: string, fileName: string) {
+      if (memoryCache.contents[fileName] !== code) {
+        memoryCache.contents[fileName] = code
+        memoryCache.versions[fileName] = (memoryCache.versions[fileName] || 0) + 1
       }
     }
 
     // Create the compiler host for type checking.
     const serviceHost = {
-      getScriptFileNames: () => Object.keys(cache.versions),
+      getScriptFileNames: () => Object.keys(memoryCache.versions),
       getScriptVersion: (fileName: string) => {
-        const version = cache.versions[fileName]
+        const version = memoryCache.versions[fileName]
 
         // We need to return `undefined` and not a string here because TypeScript will use
         // `getScriptVersion` and compare against their own version - which can be `undefined`.
@@ -299,18 +284,16 @@ export function register (options: Options = {}): Register {
         return version === undefined ? undefined as any as string : String(version)
       },
       getScriptSnapshot (fileName: string) {
-        if (!cache.contents[fileName]) {
-          if (!fileExists(fileName)) {
-            return undefined
-          }
-
-          cache.contents[fileName] = getFile(fileName)
+        if (!memoryCache.contents[fileName]) {
+          const contents = readFile(fileName)
+          if (!contents) return
+          memoryCache.contents[fileName] = contents
         }
 
-        return ts.ScriptSnapshot.fromString(cache.contents[fileName])
+        return ts.ScriptSnapshot.fromString(memoryCache.contents[fileName])
       },
       fileExists: debugFn('fileExists', fileExists),
-      readFile: debugFn('getFile', getFile),
+      readFile: debugFn('getFile', readFile),
       readDirectory: debugFn('readDirectory', ts.sys.readDirectory),
       getDirectories: debugFn('getDirectories', ts.sys.getDirectories),
       directoryExists: debugFn('directoryExists', ts.sys.directoryExists),
@@ -318,12 +301,15 @@ export function register (options: Options = {}): Register {
       getCurrentDirectory: () => cwd,
       getCompilationSettings: () => config.options,
       getDefaultLibFileName: () => ts.getDefaultLibFilePath(config.options),
-      getCustomTransformers: () => options.transformers
+      getCustomTransformers: () => transformers
     }
 
     const service = ts.createLanguageService(serviceHost)
 
-    getOutput = function (_code: string, fileName: string, lineOffset: number = 0) {
+    getOutput = function (code: string, fileName: string, lineOffset: number = 0) {
+      // Must set memory cache before attempting to read file.
+      updateMemoryCache(code, fileName)
+
       const output = service.getEmitOutput(fileName)
 
       // Get the relevant diagnostics - this is 3x faster than `getPreEmitDiagnostics`.
@@ -331,7 +317,7 @@ export function register (options: Options = {}): Register {
         .concat(service.getSyntacticDiagnostics(fileName))
         .concat(service.getSemanticDiagnostics(fileName))
 
-      const diagnosticList = filterDiagnostics(diagnostics, ignoreWarnings)
+      const diagnosticList = filterDiagnostics(diagnostics, ignoreDiagnostics)
 
       if (diagnosticList.length) {
         throw new TSError(formatDiagnostics(diagnosticList, cwd, ts, lineOffset))
@@ -355,21 +341,8 @@ export function register (options: Options = {}): Register {
       return [output.outputFiles[1].text, output.outputFiles[0].text]
     }
 
-    compile = readThrough(
-      cachedir,
-      shouldCache,
-      getFile,
-      cache,
-      function (code: string, fileName: string, lineOffset?: number) {
-        setCache(code, fileName)
-
-        return getOutput(code, fileName, lineOffset)
-      },
-      getExtension
-    )
-
     getTypeInfo = function (code: string, fileName: string, position: number) {
-      setCache(code, fileName)
+      updateMemoryCache(code, fileName)
 
       const info = service.getQuickInfoAtPosition(fileName, position)
       const name = ts.displayPartsToString(info ? info.displayParts : [])
@@ -379,6 +352,7 @@ export function register (options: Options = {}): Register {
     }
   }
 
+  const compile = readThrough(cachedir, options.cache === true, memoryCache, getOutput, getExtension)
   const register: Register = { cwd, compile, getTypeInfo, extensions, cachedir, ts }
 
   // Register the extensions.
@@ -427,9 +401,9 @@ function registerExtension (
 }
 
 /**
- * Do post-processing on config options to correct them.
+ * Do post-processing on config options to support `ts-node`.
  */
-function fixConfig (config: any, ts: TSCommon) {
+function fixConfig (ts: TSCommon, config: any) {
   // Delete options that *should not* be passed through.
   delete config.options.out
   delete config.options.outFile
@@ -451,31 +425,39 @@ function fixConfig (config: any, ts: TSCommon) {
 /**
  * Load TypeScript configuration.
  */
-function readConfig (compilerOptions: any, project: string | boolean | undefined, cwd: string, ts: TSCommon) {
-  const result = loadSync(cwd, typeof project === 'string' ? project : undefined)
+function readConfig (
+  cwd: string,
+  ts: TSCommon,
+  compilerOptions: any,
+  fileExists: (path: string) => boolean,
+  readFile: (path: string) => string | undefined,
+  project?: string | null,
+  noProject?: boolean | null
+) {
+  let config = { compilerOptions: {} as any }
+  let basePath = cwd
+  let configFileName: string | undefined = undefined
 
-  // Override default configuration options.
-  result.config.compilerOptions = Object.assign({}, result.config.compilerOptions, compilerOptions, {
-    sourceMap: true,
-    inlineSourceMap: false,
-    inlineSources: true,
-    declaration: false,
-    noEmit: false,
-    outDir: '$$ts-node$$'
-  })
+  // Read project configuration when available.
+  if (!noProject) {
+    configFileName = project ? resolve(cwd, project) : ts.findConfigFile(cwd, fileExists)
 
-  const configPath = result.path && normalizeSlashes(result.path)
-  const basePath = configPath ? dirname(configPath) : normalizeSlashes(cwd)
+    if (configFileName) {
+      const result = ts.readConfigFile(configFileName, readFile)
 
-  if (typeof ts.parseConfigFile === 'function') {
-    return fixConfig(ts.parseConfigFile(result.config, ts.sys, basePath), ts)
+      if (result.error) {
+        throw new TSError([formatDiagnostic(result.error, cwd, ts, 0)])
+      }
+
+      config = result.config
+      basePath = normalizeSlashes(dirname(configFileName))
+    }
   }
 
-  if (typeof ts.parseJsonConfigFileContent === 'function') {
-    return fixConfig(ts.parseJsonConfigFileContent(result.config, ts.sys, basePath, undefined, configPath), ts)
-  }
+  // Override default configuration options `ts-node` requires.
+  config.compilerOptions = Object.assign({}, config.compilerOptions, compilerOptions, DEFAULT_COMPILER_OPTIONS)
 
-  throw new TypeError('Could not find a compatible `parseConfigFile` function')
+  return fixConfig(ts, ts.parseJsonConfigFileContent(config, ts.sys, basePath, undefined, configFileName))
 }
 
 /**
@@ -489,8 +471,7 @@ type SourceOutput = [string, string]
 function readThrough (
   cachedir: string,
   shouldCache: boolean,
-  getFile: (fileName: string) => string,
-  cache: Cache,
+  memoryCache: MemoryCache,
   compile: (code: string, fileName: string, lineOffset?: number) => SourceOutput,
   getExtension: (fileName: string) => string
 ) {
@@ -501,7 +482,7 @@ function readThrough (
       const [value, sourceMap] = compile(code, fileName, lineOffset)
       const output = updateOutput(value, fileName, sourceMap, getExtension)
 
-      cache.outputs[fileName] = output
+      memoryCache.outputs[fileName] = output
 
       return output
     }
@@ -518,9 +499,9 @@ function readThrough (
     const outputPath = `${cachePath}${extension}`
 
     try {
-      const output = getFile(outputPath)
+      const output = readFileSync(outputPath, 'utf8')
       if (isValidCacheContent(output)) {
-        cache.outputs[fileName] = output
+        memoryCache.outputs[fileName] = output
         return output
       }
     } catch (err) {/* Ignore. */}
@@ -528,7 +509,7 @@ function readThrough (
     const [value, sourceMap] = compile(code, fileName, lineOffset)
     const output = updateOutput(value, fileName, sourceMap, getExtension)
 
-    cache.outputs[fileName] = output
+    memoryCache.outputs[fileName] = output
     writeFileSync(outputPath, output)
 
     return output
@@ -543,7 +524,7 @@ function updateOutput (outputText: string, fileName: string, sourceMap: string, 
   const sourceMapContent = `data:application/json;charset=utf-8;base64,${base64Map}`
   const sourceMapLength = `${basename(fileName)}.map`.length + (getExtension(fileName).length - extname(fileName).length)
 
-  return outputText.slice(0, -1 * sourceMapLength) + sourceMapContent
+  return outputText.slice(0, -sourceMapLength) + sourceMapContent
 }
 
 /**
@@ -572,35 +553,15 @@ function getCacheName (sourceCode: string, fileName: string) {
  * Ensure the given cached content is valid by sniffing for a base64 encoded '}'
  * at the end of the content, which should exist if there is a valid sourceMap present.
  */
-function isValidCacheContent (content: string) {
-  return /(?:9|0=|Q==)$/.test(content.slice(-3))
+function isValidCacheContent (contents: string) {
+  return /(?:9|0=|Q==)$/.test(contents.slice(-3))
 }
 
 /**
  * Create a hash of the current configuration.
  */
-function getCompilerDigest (opts: any) {
-  return crypto.createHash('sha256').update(JSON.stringify(opts), 'utf8').digest('hex')
-}
-
-/**
- * Check if the file exists.
- */
-export function fileExists (fileName: string): boolean {
-  try {
-    const stats = statSync(fileName)
-
-    return stats.isFile() || stats.isFIFO()
-  } catch (err) {
-    return false
-  }
-}
-
-/**
- * Get the file from the file system.
- */
-export function getFile (fileName: string): string {
-  return readFileSync(fileName, 'utf8')
+function getCompilerDigest (obj: object) {
+  return crypto.createHash('sha256').update(JSON.stringify(obj), 'utf8').digest('hex')
 }
 
 /**
