+++ tests/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
-    "jsx": "react"
+    "jsx": "react",
+    "noEmit": true
   }
 }
