--- typings.json
@@ -1,26 +1,0 @@
-{
-  "dependencies": {
-    "arrify": "registry:npm/arrify#1.0.0+20160723033700",
-    "chalk": "registry:npm/chalk#1.0.0+20160211003958",
-    "diff": "registry:npm/diff#2.0.0+20160723033700",
-    "make-error": "registry:npm/make-error#1.0.0+20160211003958",
-    "minimist": "registry:npm/minimist#1.0.0+20160229232932",
-    "mkdirp": "registry:npm/mkdirp#0.5.0+20160222053049",
-    "proxyquire": "registry:npm/proxyquire#1.0.0+20160211003958",
-    "source-map-support": "registry:npm/source-map-support#0.3.0+20161006200546",
-    "tsconfig": "npm:tsconfig",
-    "v8flags": "registry:npm/v8flags#3.0.0+20161011212142",
-    "xtend": "registry:npm/xtend#4.0.0+20160211003958",
-    "yn": "registry:npm/yn#1.3.0+20170508185912"
-  },
-  "globalDependencies": {
-    "node": "registry:env/node#6.0.0+20170213133316"
-  },
-  "globalDevDependencies": {
-    "chai": "registry:dt/chai#3.4.0+20160216071402",
-    "mocha": "registry:dt/mocha#2.2.5+20161028141524"
-  },
-  "devDependencies": {
-    "semver": "registry:npm/semver#5.0.0+20160211003958"
-  }
-}
